requirements.txt
flask
requests
python-dotenv
twilio
click
rich
tabulate
gunicorn
psycopg2-binary

